Refactoring global page objects
