+++ PageObjects/page_objects.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 from Components.auth_form import AuthForm
-from Components.feed_page_components import TopMenuOnShopPage, LeftMenuOnShopFeedPage, HeaderOnShopFeedPage
-from Components.forum_page_components import TopicPopup
-from Components.groups_page_components import CreateGroupsPopup
+from Components.shop_feed_page_components import TopMenuOnShopPage, LeftMenuOnShopFeedPage, HeaderOnShopFeedPage
+from Components.shop_forum_page_components import TopicPopup
+from Components.groups_page_components import CreateGroupPopup
 from Components.main_page_components import LeftMenuOnMainPage
-from Components.market_page_components import CatalogPopup, CatalogWidget, CatalogPanel, RemoveCatalogPopup, \
+from Components.shop_market_page_components import CatalogPopup, CatalogWidget, CatalogPanel, RemoveCatalogPopup, \
     CatalogCounter, ProductCounter, CatalogStub, CreateProductPopup, ProductWidget, ProductStub, \
     SubmitProductActionPopup
 from PageObjects.page import Page
@@ -25,7 +25,7 @@ class MainPage(Page):
 class GroupsPage(Page):
     @property
     def popup(self):
-        return CreateGroupsPopup(self.driver)
+        return CreateGroupPopup(self.driver)
 
 
 class ShopFeedPage(Page):
